Move digest types to a new file
